refactor: extracts promote/demote/metadata/deploy logic into flows
