+++ scripts/deploy.py
@@ -1,4 +1,10 @@
-"""The whole pipeline to deploy classifier models."""
+"""
+CLI wrapper for the classifier deployment pipeline.
+
+The reusable train -> promote -> refresh logic lives in `flows.deploy`; this module
+only adds the Typer commands used by `just deploy-existing` / `just deploy-new`, plus
+the CLI-level argument validation and exit handling.
+"""
 
 import traceback
 from pathlib import Path
@@ -7,23 +13,31 @@ from typing import Annotated
 import typer
 from dotenv import load_dotenv
 
-import scripts.promote
-import scripts.train
-from knowledge_graph.cloud import (
-    AwsEnv,
-    parse_aws_env,
-    parse_spec_file,
-    validate_transition,
+from flows.deploy import (
+    run_deploy_existing,
+    run_deploy_new,
+    validate_classifiers_profiles,
 )
+from knowledge_graph.cloud import AwsEnv, parse_aws_env
 from knowledge_graph.identifiers import WikibaseID
-from scripts.update_classifier_spec import (
-    refresh_all_available_classifiers,
-)
 
 # Load environment variables from .env file at Git root
 load_dotenv(Path(__file__).parent.parent / ".env")
 
 app = typer.Typer()
+
+
+def _validate_classifiers_profiles(
+    add_classifiers_profiles: list[str] | None,
+    remove_classifiers_profiles: list[str] | None = None,
+) -> None:
+    """Validate profiles at the CLI layer, surfacing errors as `typer.BadParameter`."""
+    try:
+        validate_classifiers_profiles(
+            add_classifiers_profiles, remove_classifiers_profiles
+        )
+    except ValueError as e:
+        raise typer.BadParameter(str(e))
 
 
 @app.command()
@@ -54,51 +68,18 @@ def existing(
     ] = None,
 ):
     """Deploy existing models from one environment to another."""
-    validate_transition(from_aws_env, to_aws_env)
-
-    add_class_prof: set[str] = (
-        set(add_classifiers_profiles) if add_classifiers_profiles else set()
+    _validate_classifiers_profiles(
+        add_classifiers_profiles, remove_classifiers_profiles
     )
-    remove_class_prof: set[str] = (
-        set(remove_classifiers_profiles) if remove_classifiers_profiles else set()
+
+    run_deploy_existing(
+        from_aws_env=from_aws_env,
+        to_aws_env=to_aws_env,
+        train=train,
+        promote=promote,
+        add_classifiers_profiles=add_classifiers_profiles,
+        remove_classifiers_profiles=remove_classifiers_profiles,
     )
-    if dupes := add_class_prof & remove_class_prof:
-        raise typer.BadParameter(
-            f"duplicate values found for adding and removing classifiers profiles: `{','.join(dupes)}`"
-        )
-
-    if len(add_class_prof) > 1:
-        raise typer.BadParameter(
-            f"Artifact must have maximum of one classifiers profile in metadata, or you must specify 1 to remove. Provided: `{','.join(add_class_prof)}`"
-        )
-
-    specs = parse_spec_file(from_aws_env)
-    print(f"loaded {len(specs)} classifier specifications")
-
-    for spec in specs:
-        print(f"\nprocessing {spec.name}:{spec.alias}")
-
-        if train:
-            print("training")
-            classifier = scripts.train.main(
-                wikibase_id=WikibaseID(spec.name),
-                track_and_upload=True,
-                aws_env=to_aws_env,
-            )
-            if not classifier:
-                raise ValueError("No classifier returned from training.")
-
-            if promote:
-                print("promoting")
-                scripts.promote.main(
-                    wikibase_id=WikibaseID(spec.name),
-                    classifier_id=classifier.id,
-                    aws_env=to_aws_env,
-                    add_classifiers_profiles=add_classifiers_profiles,
-                    remove_classifiers_profiles=remove_classifiers_profiles,
-                )
-
-    refresh_all_available_classifiers([to_aws_env])
 
 
 @app.command()
@@ -126,47 +107,15 @@ def new(
     ] = None,
 ):
     """Deploy new models by training and promoting them."""
-    failed_wikibase_ids = []
-    for wikibase_id in wikibase_ids:
-        try:
-            print(f"\nprocessing {wikibase_id}")
+    _validate_classifiers_profiles(add_classifiers_profiles)
 
-            add_class_prof: set[str] = (
-                set(add_classifiers_profiles) if add_classifiers_profiles else set()
-            )
-
-            if len(add_class_prof) > 1:
-                raise typer.BadParameter(
-                    f"Artifact must have maximum of one classifiers profile in metadata. Provided: `{','.join(add_class_prof)}`"
-                )
-
-            if train:
-                print("training")
-                classifier = scripts.train.main(
-                    wikibase_id=wikibase_id,
-                    track_and_upload=True,
-                    aws_env=aws_env,
-                )
-
-                if not classifier:
-                    raise ValueError("No classifier returned from training.")
-
-                if promote:
-                    print("promoting")
-                    scripts.promote.main(
-                        wikibase_id=wikibase_id,
-                        classifier_id=classifier.id,
-                        aws_env=aws_env,
-                        add_classifiers_profiles=add_classifiers_profiles,
-                    )
-        except AttributeError as e:
-            print(f"Error getting concept: {e}")
-            failed_wikibase_ids.append((wikibase_id, e))
-            continue
-
-    refresh_all_available_classifiers([aws_env])
-
-    if failed_wikibase_ids:
+    if failed_wikibase_ids := run_deploy_new(
+        aws_env=aws_env,
+        wikibase_ids=wikibase_ids,
+        train=train,
+        promote=promote,
+        add_classifiers_profiles=add_classifiers_profiles,
+    ):
         for wikibase_id, e in failed_wikibase_ids:
             print(f"Failed to deploy classifier for {wikibase_id}:")
             print("".join(traceback.format_exception(e)))
